% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultraread.R
\name{ultraread}
\alias{ultraread}
\title{ultraread}
\usage{
ultraread(
  filepath = NULL,
  type = "csv",
  sheetnum = 1,
  headed = TRUE,
  encoded = "UTF-8",
  driver = "ODBC Driver 17 for SQL Server",
  server = "nameof.database.windows.net",
  database = "nameofdatabase",
  uid = "database_username",
  pwd = "databasepassword",
  port = 1433
)
}
\arguments{
\item{filepath}{Required for csv and xlsx file reading}

\item{type}{Defaults to "csv", also accepts "xlsx" filetype or "db" to get started with a database connection}

\item{sheetnum}{Defaults to 1, preferably use the index number of the sheet being returned as a dataframe. Only relevant for xlsx files}

\item{headed}{Defaults to TRUE, turns the first row into the column names of the returned dataframe}

\item{encoded}{Defaults to "UTF-8". The encoding you require for reading your file. Only relevant to CSVs}

\item{driver}{Defaults to "ODBC Driver 17 for SQL Server". The DBI::dbConnect compatible driver you wish to use. The driver must be installed correctly on your computer. See https://learn.microsoft.com/en-us/sql/connect/odbc/download-odbc-driver-for-sql-server?view=sql-server-ver16}

\item{server}{The database server you wish to connect to}

\item{database}{The name of the database you wish to connect to}

\item{uid}{Your username or user-identification for logging into the database}

\item{pwd}{Your password for logging into the database}

\item{port}{Defaults to 1433. The port you wish to connect to the database via}
}
\value{
Your csv or xlsx sheet as a dataframe, otherwise a database connection.
}
\description{
Create Dataframes from CSV, XLSX or DBI connections
}
\note{
Will result in error when filepath is given but csv or xlsx types are specified. For database connections, the ODBC driver you specify must be installed on the computer you are using. It also must have established a connection to the database. For more info see https://learn.microsoft.com/en-us/sql/connect/odbc/download-odbc-driver-for-sql-server?view=sql-server-ver16
}
\examples{
## Not run:
df = ultraread(filepath = "./encoding test.xlsx", type = "xlsx", sheetnum = 2, headed = TRUE)
df = ultraread(filepath = "./encoding test.csv", type = "csv", headed = TRUE, encoded = "latin1")
con = ultraread(server = "company.database.windows.net", database = "sales_db", uid = "user_x", pwd = "secret_password", port = 1433)
##End(Not run)
}
